+++ changelog.md
@@ -1,7 +1,9 @@
 ## version 1.3.0 [In Progress]
 1. Deprecated: `getdoc()`;
 2. Feature: support app['lessweb.routes'];
-3. Feature: Support multiple positional-only parameters;  //TODO: need testing
+3. Feature: support multiple positional-only parameters;
+4. Feature: add e2e test (new tests/e2e/ dir);
+5. Feature: when incorrectly used, raise TypeError as the standard specification.
 
 ## version 1.2.0
 1. Feature: support typecasting Union and NewType;